# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Compras por Aplicativo",
        "description": "Compras Por Aplicativo - Flutter/NodeJS",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/thinnugly",
            "email": "renatomuiambo24@gmail.com"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/slider": {
            "get": {
                "tags": [
                    "Slider"
                ],
                "summary": "Get All Slider",
                "parameters": [
                    {
                        "name": "sliderName",
                        "in": "query",
                        "required": false,
                        "description": "Slider Name",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "Page Number",
                        "type": "integer",
                        "default": 1
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "description": "Page Number",
                        "type": "integer",
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SliderServiceResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Slider"
                ],
                "summary": "Create Slider",
                "parameters": [
                    {
                        "name": "sliderName",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "Slider Name (unique)"
                    },
                    {
                        "name": "sliderDescription",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "Slider Description"
                    },
                    {
                        "name": "sliderImage",
                        "in": "formData",
                        "required": false,
                        "type": "file",
                        "description": "Slider Image"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SliderServiceResponse"
                        }
                    }
                }
            }
        },
        "/api/slider/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Slider Id",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Slider"
                ],
                "summary": "Slider Details based on Slider Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SliderServiceResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Slider"
                ],
                "summary": "Update Slider",
                "parameters": [
                    {
                        "name": "sliderName",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "Slider Name (unique)"
                    },
                    {
                        "name": "sliderDescription",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "Slider Description"
                    },
                    {
                        "name": "sliderImage",
                        "in": "formData",
                        "required": false,
                        "type": "file",
                        "description": "Slider Image"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SliderServiceResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Slider"
                ],
                "summary": "Delete Slider",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SliderServiceResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Slider": {
            "required": [
                "sliderName",
                "sliderImage",
                "_id"
            ],
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "sliderName": {
                    "type": "string"
                },
                "sliderDescription": {
                    "type": "string"
                },
                "sliderURL": {
                    "type": "string"
                },
                "sliderImage": {
                    "type": "string"
                }
            }
        },
...
        "SliderServiceResponse": {
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Slider"
                    }
                }
            }
        }
    }
}